requirements.txt
# Custom built packages
my-utils-library

# Project utils dependencies
python-dotenv
